refactor: move string utility functions to utils_convert.c for better organization
